requirements.txt
asgiref==3.11.0
brotli==1.2.0
cffi==2.0.0
cssselect2==0.8.0
Django==5.2.9
django-phonenumber-field==8.4.0
django-pwa==2.0.1
fonttools==4.61.1
phonenumbers==9.0.21
pillow==12.0.0
pycparser==2.23
pydyf==0.12.1
pyphen==0.17.2
sqlparse==0.5.4
tinycss2==1.5.1
tinyhtml5==2.0.0
tzdata==2025.3
weasyprint==67.0
webencodings==0.5.1
zopfli==0.4.0
